Full implementation without conflict tests and fixes

+++ transaction_daemon/app/managers/filesystem_manager.py
@@ -3,7 +3,14 @@ import os
 from ..utils import exceptions as ex
 from ..utils import logging
 
-# It is assumed that the root folder "/" and specifically "/tmp" is not (part of) a zfs file system
+# =================================================================================================
+# Class: FilesystemManager
+# -------------------------------------------------------------------------------------------------
+# Provides a number of methods to interact with the file system, zfs datasets and zfs snapshots
+# Locks all managed filesystems/zfs pools, such that only sudo can directly manipulate files
+#  without the use of this program
+# It is assumed that the root folder "/" and specifically "/tmp" are not (part of) a zfs file system
+# =================================================================================================
 class FilesystemManager:
     _instance = None
 
@@ -45,21 +52,16 @@ class FilesystemManager:
             cls._instance = super().__new__(cls)
             print(f"[FM][INFO] new instance {cls._instance} created")
         return cls._instance
-
-    def get_instance(self):
-        if cls._instance == None: cls._instance = FilesystemManager()
-        return cls._instance
 # =============================================================================
-# ZFS Interaction
+# ZFS Dataset Interaction
 # =============================================================================
     # Method: get_fs ----------------------------------------------------------
-    # Takes in a path and returns the corresponding filesystem. creates a new filesystem if it doesn't exist yet
+    # Takes in a path and returns the corresponding filesystem.
+    # Raises an exception if the path does not exist as a filesystem
     def get_fs(self,path,is_file=True):
-        print(f"Path before: {path}")
-        if not is_file:
-            path = f'{path}{os.sep}'
+        if not is_file: 
+            path = f'{path}{os.sep}' # Add Separator to force 'path' to be interpreted as having no file
         fs_name = os.path.dirname(path).strip(os.sep)
-        print(f"fs_name: {fs_name}")
         if fs_name not in self._filesystems: raise ex.FilesystemNotFoundException
         return fs_name
     # Method: make_fs ---------------------------------------------------------
@@ -68,24 +70,29 @@ class FilesystemManager:
     # when path = "/tank/file.txt" ('file.txt' will be a folder!)
     def make_fs(self,path):
         fs_name = path.strip(os.sep)
-        print("fs_name: {fs_name}")
-        if fs_name in self._filesystems: return 0
+        if fs_name in self._filesystems: raise ex.FilesystemExistsException
         if not any(top_level_fs in fs_name for top_level_fs in self._top_level_fs): 
             raise ex.TopLevelFsNotFoundException
         subprocess.run(["sudo","zfs","create",fs_name])
-        print("filesystem created.")
         self._filesystems.append(fs_name)
         return 0
-    # Method: destroy_fs -----------------------------------------------------
+    # Method: destroy_fs ------------------------------------------------------
     # Takes a path and destroys the corresponding filesystem, if it exists
-    # [WARNING] All files will be lost, use with caution!
+    # If a path has subpaths or subdirectories only the files in the given path will
+    # be destroyed (i.e.: if tank/foo is destroyed, but tank/foo/bar exists, the second will be preserved
+    # [WARNING] All files in dataset will be lost, use with caution!
     def destroy_fs(self,path):
-        pass
-
-
-    # Snapshots --------------------------------------------------------------
+        fs_name = path.strip(os.sep)
+        if fs_name not in self._filesystems: return 0
+        subprocess.run(["sudo", "zfs", "destroy", "-r", fs_name])
+        self._filesystems.remove(fs_name)
+        return 0
+# =============================================================================
+# ZFS Snapshot Interaction
+# =============================================================================
     def create_snapshot(self,path,id):
-        fs_name = self.get_fs(path)
+        fs_name = self.get_fs(path, is_file=False)
+        if fs_name not in self._filesystems: raise ex.PathNotFoundException
         snap_name = f'{fs_name}@{id}'
         subprocess.run(["sudo", "zfs", "snapshot", snap_name])
         self.logger.log(f"[FM][INFO] Snapshot {snap_name} created")
@@ -94,31 +101,51 @@ class FilesystemManager:
     def restore_snapshot(self, snap_name):
         subprocess.run(["sudo","zfs","rollback","-r",snap_name])
     def destroy_snapshot(self,snap_name):
+        if snap_name is None: return
         subprocess.run(["sudo","zfs","destroy",snap_name])
         self.logger.log(f"[FM][INFO] Snapshot {snap_name} destroyed")
 # =============================================================================
 # File Interaction
 # =============================================================================
+    # Method: create_file_copy ------------------------------------------------
+    # [EX] FileNotFoundError, ex.FilesystemNotFoundException, ex.NotAFileException
     def create_file_copy(self, origin_path, id):
+        self.get_fs(origin_path) # Crude way to check if the path actually exists.
+        if not os.path.exists(origin_path): raise FileNotFoundError
+        if not os.path.isfile(origin_path): raise ex.NotAFileException
         _, filename = os.path.split(origin_path)
         copy_filename = f'{id}_{filename}'
         copy_path = f'{self._filecopy_path}/{copy_filename}'
-        subprocess.run(["touch", copy_path])
-        subprocess.run(["chmod", "666", copy_path])
+        subprocess.run(["sudo", "touch", copy_path])
+        subprocess.run(["sudo", "chmod", "666", copy_path])
         return copy_path
-    # TODO
-    def create_file(self,path):
-        dirname, _ = os.path.split(path)
-        fs_name = self.get_fs(dirname)
-        if not fs_name in self._filesystems: raise PathNotFoundException
-        subprocess.run(["sudo", "touch", path])
-
-    def delete_file(self,path):
-        if not os.path.exists(path): raise ex.PathNotFoundException
-        if not os.path.isfile(path): raise ex.NotAFileException
-        subprocess.run(["sudo","rm", path])
+    def delete_file_copy(self,copy_path):
+        if not os.path.exists(copy_path): return 0
+        subprocess.run(["sudo", "rm", copy_path])
         
+    # Method: create_file -----------------------------------------------------
+    # [EX] ex.NotAFileException, ex.FilesystemNotFoundException,
+    def create_file(self,path):
+        _, filename = os.path.split(path)
+        if filename == "": raise ex.NotAFileException
+        self.get_fs(path)
+        subprocess.run(["sudo", "touch", path])
+    # Method: delete_file -----------------------------------------------------
+    # [EX] ex.NotAFileException, ex.FilesystemNotFoundException
+    def delete_file(self,path):
+        if not os.path.isfile(path): raise ex.NotAFileException
+        self.get_fs(path)
+        print("before deletion")
+        subprocess.run(["sudo","rm", path])
+    # Method: overwrite_file --------------------------------------------------
     def overwrite_file(self, origin_path, copy_path):
-        subprocess.run(["sudo", "mv", copy_path, origin_path])
-        subprocess.run(["sudo", "chmod", "644", origin_path])
-        self.logger.log(f"[FM][INFO] {copy_path} -> {origin_path}")
+        try:
+            self.get_fs(origin_path)
+            subprocess.run(["sudo", "mv", copy_path, origin_path])
+            subprocess.run(["sudo", "chmod", "644", origin_path])
+            self.logger.log(f"[FM][INFO] {copy_path} -> {origin_path}")
+        except Exception:
+            self.delete_file(copy_path)
+
+
+
